requirements.txt
aiohttp
aiofiles
httpx
playwright
openai
moviepy
ffmpeg-python